feat: ✨ Release v0.2.1
