+++ src/transaction/stream.rs
@@ -1,0 +1,43 @@
+pub fn default_hex() -> String {
+    "0x".to_string()
+}
+
+pub fn default_fee_per_byte() -> u64 {
+    8
+}
+
+pub fn default_max_block_size() -> u32 {
+    4
+}
+
+pub fn default_max_txn_size() -> u32 {
+    8
+}
+
+pub fn default_overall_burn_percentage() -> u32 {
+    4
+}
+
+pub fn default_reward_per_year() -> u64 {
+    8
+}
+
+pub fn default_validator_count_limit() -> u32 {
+    4
+}
+
+pub fn default_joining_fee() -> u64 {
+    8
+}
+
+pub fn default_claiming_fee() -> u64 {
+    8
+}
+
+pub fn default_fee_share() -> u64 {
+    8
+}
+
+pub fn default_proposal_status() -> String {
+    "ongoing".to_string()
+}
