+++ src/validator/mod.rs
@@ -1,0 +1,16 @@
+use std::net::IpAddr;
+use serde::{Deserialize};
+
+#[derive(Debug, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct Validator {
+    pub address: String,
+    pub ip: IpAddr,
+    #[serde(default)]
+    pub bad_actor: bool,
+    pub voting_power: u64,
+    pub total_shares: u64,
+    pub delegators_count: u32,
+    #[serde(default)]
+    pub is_active: bool,
+}
